Remove wgpu shader, rendering clean up (#310)
