+++ compositor_render/src/transformations/shader/base_params.rs
@@ -4,13 +4,13 @@ use crate::Resolution;
 
 #[repr(C)]
 #[derive(Debug, bytemuck::Pod, bytemuck::Zeroable, Clone, Copy)]
-pub struct CommonShaderParameters {
+pub struct BaseShaderParameters {
     time: f32,
     pub texture_count: u32,
     output_resolution: [u32; 2],
 }
 
-impl CommonShaderParameters {
+impl BaseShaderParameters {
     pub fn new(time: Duration, texture_count: u32, output_resolution: Resolution) -> Self {
         Self {
             time: time.as_secs_f32(),
@@ -23,7 +23,7 @@ impl CommonShaderParameters {
     }
 
     pub fn push_constant_size() -> u32 {
-        let size = std::mem::size_of::<CommonShaderParameters>() as u32;
+        let size = std::mem::size_of::<BaseShaderParameters>() as u32;
         match size % 4 {
             0 => size,
             rest => size + (4 - rest),
